Отправка формы перенесена в отдельный модуль
