+++ js/main.js
@@ -4,3 +4,4 @@ import './html-ad-generator.js';
 import './util.js';
 import './map.js';
 import './similar-ad-pins-renderer.js';
+import './form-data-sender.js';
